+++ lib/brando_admin/components/layouts.ex
@@ -1,0 +1,4 @@
+defmodule BrandoAdmin.Layouts do
+  use BrandoAdmin, :html
+  embed_templates "layouts/*"
+end