+++ lib/brando_admin/views/admin_view.ex
@@ -1,7 +1,0 @@
-defmodule BrandoAdmin.AdminView do
-  @moduledoc """
-  Main view for Brando admin.
-  """
-  use BrandoAdmin, :view
-  alias BrandoAdmin.Components.Content
-end