+++ priv/templates/brando.install/lib/application_name_web/controllers/fallback_controller.ex
@@ -1,11 +1,11 @@
 defmodule <%= application_module %>Web.FallbackController do
-  use <%= application_module %>Web, :controller
+  use BrandoWeb, :controller
 
   def call(conn, {:error, {_, :not_found}}) do
     conn
     |> put_status(:not_found)
     |> put_layout(false)
-    |> put_view(<%= application_module %>Web.ErrorView)
+    |> put_view(<%= application_module %>Web.ErrorHTML)
     |> render(:"404")
   end
 
@@ -13,7 +13,7 @@ defmodule <%= application_module %>Web.FallbackController do
     conn
     |> put_status(:unauthorized)
     |> put_layout(false)
-    |> put_view(<%= application_module %>Web.ErrorView)
+    |> put_view(<%= application_module %>Web.ErrorHTML)
     |> render(:"401")
   end
 end