+++ priv/templates/brando.install/lib/application_name_web/views/lockdown_view.ex
@@ -1,3 +1,0 @@
-defmodule <%= application_module %>Web.LockdownView do
-  use <%= application_module %>Web, :view
-end